Refactor 'process' entity into its own file
